+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Added
 - Add support for bitmasks to `Spatial` component and raster schema.
 - Add loader for `GeoJSON` polygons/points as cells.  QuPath outputs GeoJSON for annotations.
+- Updated `GeoJSON` loader to support the Baysor custom format
+- Added coordination types for selection of molecules
 
 ### Changed
 - Use GH Action for Cypress specifically due to random failures on OME-TIFF example.